r3.in.v5d, r3.out.v5d: Deprecate CRAY HW specific code (#4545)
Currently, there is no sign either from GitHub issues or grass-dev archive that many people are using CRAY code which was introduced 25 years ago. As part of fixing a bug, we found that it was hard to find reference materials related to CRAY, due to it being old and not in usage anymore. For this reason, deprecate CRAY code and corresponding complexity that comes with it.

Signed-off-by: Mohan Yelugoti <ymdatta.work@gmail.com>