lib/vector/Vlib: Initialize variable before using it in write_sfa module
In `V2__add_line_to_topo_sfa` function which is only compiled when
`HAVE_OGR` or `HAVE_POSTGRES` is defined, `first` variable is used
without being initialized, which is undefined behavior.

Based on the context, initialize `first` to `TRUE`, so that we first
copy and then extend a box.

This was found using cppcheck tool.

Signed-off-by: Mohan Yelugoti <ymdatta.work@gmail.com>