gui: Fix possibly unbound variables in psmap.utils.BBoxAfterRotation (#4873)
Current code didn't handle the case where angle is 0, nor if out of range (more than 360 degrees). For example, calling with an angle of 375.7 will try to compute the width and height on variables not set yet. This handles these two cases, and short-circuits for an angle of 0 degrees.

Angles greater than 360 degrees can be supported by this, but range documented is kept as is.