r.mapcalc: Use mask and not MASK in the doc
Uses mask or raster mask instead of MASK in the documentation. Where the raster called MASK needs to be accessed directly in an example, MASK is used as the assumed name mentioning that explicitly.

Updates r3.mapcalc in the same way which should be harmless since the wording does not explicitly mention the possibility of a different name. The text looks like a direct copy-paste from r.mapcalc, so that is preserved. The only improvement is that r3.mask is now linked instead of r.mask (in the 3D doc).