d.rast.edit: Use raster mask, not uppercase MASK (#5001)
Replaces usage of uppercase MASK by 'raster mask' in the documentation. This clarifies the sentence in terms of what respecting means (original says 'respected...including any edits' which may suggest the opposite). It gives me too many errors to verify the behavior as the comments asks for, but the code uses r.patch, so the mask is likely applied and thus the mask is applied for any writing (unlike r.mapcalc).