wxGUI: Use r.mask.status to get info about mask (#4519)
This replaces low level checks of MASK raster in the current mapset with calls of r.mask.status.

GConsole and SbMask require r.mask.status to output a potential name of the mask regerdless of mask being set or not, so this functionality still needs to be implemented first before moving this PR forward. History changes seem to be complete.