v.to.rast: Give verbose message about number for all types (#2498)
When there are no areas in the vector map, v.to.rast gave warning, and now after #2459, it gives a verbose message about number of selected areas when areas are selected as type. This gives the verbose message for all types when they are selected making the behavior more consistent across types (points, lines, boundaries, and centroids).

The default for type is point, line, and area, so three verbose messages are produced (when verbose output is enabled) for any vector map.