doc: Enable Markdown doc compilation
The initial commit simply uncoments the Markdown compilation lines from #3849. This requires the actual Markdown files to be present, so this breaks the CI. This will stay like this for now. The PR is mean to be tested together with the Markdown files generated by script in #4620 (switch to this branch after generating the files).

Subsequent commits will add fixes of the compilation (missing tools, files, etc.).