doc: Hide only general and misc keywords in index
When a tool links raster or vector as a standard keyword, the link to keyword page does not go anywhere because first keyword (aka class or family) is hidden in the keywords index (overview).

This enables most of the keywords with the exception of general and misc which I assume will not be used by non-general tools as keywords.

Fixes #5049.

In addition, this changes the variable names for better documentation.