doc: Convert HTML to Markdown (#5064)
This converts HTML markup in .md files to Markdown using the ./utils/grass_html2md.sh script.

The conversion was done using the #5054 version of the script with additional (reviewed) fixes by markdownlint-fix from pre-commit.

Remove .md files which should have stayed .html. This breaks history but not terribly and the history is not that important for these two (one is a recently added mkdocs template and the other may be even unused).