doc: Enable Markdown doc compilation (#5048)
The initial commit simply uncomments the Markdown compilation lines from #3849. (This requires the actual Markdown files to be present which are now part of the source code.)

Subsequent commits add fixes of the compilation (missing tools, files, etc.):

* Add compilation of multi-program directories like r.colors, r.mapcalc, and r3.flow
* Generate colortable thumbnails. This makes the script little more general and assumes that HTML will be eventually removed (the thumbnails are generated twice now, once for HTML and once of Markdown).
* Copy barscale and northarrow files for Markdown doc source.
* Presence of spaces vs underscores in key may differ between HTML and Markdown, so always do replacements.
* Do not add dash to visible text for keyword. Additionally, remove underscore just like in the HTML version.
* Add projectionintro and grass_database pages from doc dir.
* Add compilation for GUI, icons, and additional non-Markdown files.
* Add basic port of graphical index (removes broken links and provides some content).
* Strip keywords from whitespace and skip empty ones

The complex r.mapcalc build would fail on the first run attempt presumably because of a raise condition with color table build. This removes it from standard SUBDIR and relies on the extra pre-builld of r.mapcalc dir. However, this fails to do distclean properly and needs to be addressed later.

g.extension now creates md from html on the fly if no md is present. The build of Markdown and HTML doc creates a race condition while getting the metadata though git clone, so the build now checks for presence of grass-addons clone and uses the directory as a cache after a successful clone. The code assumes all directory move errors are only the good case of a cache being created in the meantime. The code also creates two (or more) clones if multiple processes run together (git clone runs twice, but the clash in the cache is avoided).

---------

Co-authored-by: Anna Petrasova <kratochanna@gmail.com>