ci: Remove Travis-CI (backport)
(Backport of #5099)

As explained in #4313, OSGeo ends its Travis subscription. Travis is not a crucial component of our CI, so we are removing it.

Additionally, travis became problematic after the Markdown documentation build was enabled because Travis limits size of the log and terminates the build if it hits the limit.