GUI: Use for Markdown documentation if available (#5066)
This uses the mkdocs-generated Markdown documentation if available and fallbacks to the original HTML doc if the built doc is not available. It assumes the old HTML doc always exists. It resolves the installed addons only to the extent the original code resolves it, so it should work as long as they come with the legacy HTML doc.