build: Reduce need for getting addons version
The corresponding addons branch, i.e., version is determined from git ls-remote which needs to go online. This limits when it is used by limiting the scope of the variables just to where they are needed. Additionally, it caches the branch (version), so the wrapper function can be used more freely on multiple places. It refactors some of the relevant common code between HTML and Markdown builds.