docs: Use meta page info from HTML in Markdown (#5169)
Special pages like raster intro or r.li in the orginal HTML doc contain one or more of meta page such as name or description.

The conversion to Markdown ignored them. This brings the info over to Markdown in YAML headers. There is no use for it at this point. They are transferred as is except `name description` which is changed to just `description`.

Additionally, `description` if present is used as a level 1 heading. This makes pages to behave like the intro pages (e.g., raster intro). This is a slight change from HTML doc where r.li would try to behave like a tool, but that's more complex and I don't see why that would be better. 

Specifically this:

- Transfers meta page from HTML to Markdown as YAML header.
- Transfers meta page description to a Markdown heading if description is present.
- Moves heading level up in several files because using heading level 1 on the source code level requires further changes in the files to avoid skipping one heading level. (That's a linting issue.)
- Modifies one database (as in project parent directory) heading text to avoid having a duplicate heading. (That's also a linting issue.)
- Modifies the build procedure to combine YAML headers in files with generated YAML headers by parser. (For all files.)