docs: Add material mkdocs tags plugin (#5153)
This uses existing keywords in Material MkDocs tags plugin to show tag. The tags mirror the keywords in terms of list and index, but are simpler, completely driven by mkdocs without any needed for custom code. The tool family and topic logic is obviously not include. More evaluation and de-duplication with keywords is needed (as a next step). 

Adds tags to parser output which are just keywords but in full YAML syntax (mkdocs consumes tags by default, but can be configured).

Uses absolute URL instead of relative links in footer to make links from addons work.

Closes #5091