docs: Replace table by sections for standard parser options (#5235)
This changes representation of the generated documentation of the standard parser options (parameters). The table did not work great in the orginal HTML documentation, but it was not useable in the new mkdoc-rendered documentation.

The options (parameters) are now represented as individual sections (level 3 headings). The individual key-value pairs for the defaults are represented as rows in a small table for each option. Examples are included to demostrate basic Python usage and allow copy-pasting.

The page now starts with a short introduction and links to g.parser for more context.

The C code parsing is now strictly split between the line ingesting and value cleanup. The cleanup is now done with regular expressions. Some additional adjustments to the values are done in the Markdown representation for clarity on the C constants and function calls.