build: Fix basic gallery build issues for Markdown (#5247)
This fixes build of header and footer by avoiding YAML syntax issue in title (colon) and writing of HTML footer. It also does multiline matching for the image capture which results in capturing (presumably) all images in the Markdown docs.

Less complete, but much faster expression. We are not capturing the alt, only testing the image presence, so we take the assumption that closing square bracket and image name in parentheses is a link. This avoids costly multiline matching with dot (from 1.5min back to original 30sec), while still capturing the multiline images (by ignoring their multiline alts).