docs: Add captions to Markdown manual gallery (#5250)
This adds simple captions to the images in the manual gallery in Markdown just like the custom HTML doc has.

Additionally, this skips small images if PIL is available on the system (but does not fail and instead uses all if PIS is not available). It tests width only because the images are shown line by line and that's what we have in the style guide. This saves about 10 sec in the build (from 30 to 20 sec).

It also sorts the images by the page name they link to which happens to put the display tools on top.