lib/parser: Separate Markdown code from reStructuredText (#5240)
This restores the original 'rest' file for reStructuredText output so that it is in a separate function (and file) from Markdown. The reStructuredText output is removed from the Markdown code. While this currently introduces some duplication, it allows for further modifications of Markdown without a need to modify the reStructuredText at the same time. Even now, there are significant differences between the outputs due to the differences in syntax (headings, images, command line option lists). Anyway, the reStructuredText is scheduled for removal since the initial addition of the Markdown build, so a separate file will be easy to remove later (and this is basically a first step in the removal).

It simplifies the Markdown code a little bit by avoiding a simple call to static function which is not needed anymore since no decision is made to switch between Markdown and reStructuredText at that level anymore.