docs: Use keywords property in MkDocs, not separate tags
Use the keywords front matter property named after keywords of a tool (and used by the custom keywords mechanism), not an additional tags property. This means that keywords property needs to be proper YAML list, not a plain comma-separated list as a string. This is how keywords are now genered. The custom keyword parsing now accepts both version (with and without []).

This does not include any keywords entered manually to Markdown because there are none in this repo, but they are in grass-addons repo which will break the MkDocs reading of the tags from the keywords property (for r.pi, r.green).

This includes cleanup of keywords tool metadata syntax which is supposed to use multiple keyword keys now rather than keywords (for a single keyword).

For the tags page generated by MkDocs, this adds an exclude of keyword 'unit test' which does not need to appear in the documentation.