docs: Link to addons repo for addons (#5319)
While #3849 added link to the specific commit, the URL was always main repo. This rewrites the code to get pure repo URL in a variable and then uses that to link to the specific commit in the appropriate repo. It limits the use of urlparse.urljoin because that is not required to join URLs and brings additional complexity with the need to add trailing slashes.