build: Add colortables and other files to addon doc (#5316)
This adds missing files to the addons subdirectory when the documentation is build in the CI. It gets all the directories in the parent dir and copies them over except the addons directory itself and libpython directory (filtering the two know unwanted ones out rather than trying to capture all directory names). An alternative solution would be to generate the links differently when compiling addons, but that would need post-processing like for keywords or changes to the parser code.