docs: Use keywords property in MkDocs, not separate tags (#5318)
In MkDocs, use the keywords front matter property keywords, named after keywords of a tool (and used by the custom keywords mechanism), not an additional tags property which had to be emitted by parser. This only works when the keywords property is a proper YAML list, not a plain comma-separated list as a string (which is true with 08286fbe4ac267017a7a39c3189b59cbe03bf2ce).