docs: Point to the repo in MkDocs GitHub stats (#5326)
The issue was caused when setting the MkDocs config variable repo_url to OSGeo org. We do this because we are combining two GitHub repos (grass, grass-addons) into one MkDocs site. New extra variables in the MKDocs config file express the full configuration and new partial source.html to points the Github stats header to OSGeo/grass instead of OSGeo. The may need further refactoring to teach the the View/Edit functionality about the variables and to avoid the partial source.html.

Fixes #5237