docs: Add sphinx-sitemap generation to Python doc (#5331)
Added sitemap for Sphinx docs using the sphinx-sitemap extension to address part of issue #5321. Includes the corresponding change in CI.

The results are a basic sitemap with each URL. Leaving out pages according to the example in documentation. Not including lastmod which is in our current site maps because this is not currently supported by Sphinx (but in it is  works).