docs: Remove old footer link from MkDocs site
This is removing the series of links in the footer of the new documentation which go to the various index pages. These links to index pages served as a menu in the custom HTML documentation, but don't have much use in the new documentation which is using MkDocs navigation with a top bar menu being the main navigation place. In very special cases, the bottom bar contained extra link to GUI and the link to tool category would be generated dynamically. This has a limited use as menus are not expected to be changing, instead you expect to find things in the same place there. (So this is something which the new MkDocs menus don't have, but it is not a problem.) In the future, if anything from this footer links should be linked from menu, it should be added to the new menu. The full index having links to all tools is part of the menu.

Closes #5314