grass.app: Use sys.executable as a GRASS_PYTHON default
Instead of hardcoding python3 and python3.exe (based on the platform), use sys.executable for the GRASS_PYTHON variable when not set externally. This switches from command-name-only to full path because sys.executable is 'absolute path  of the executable binary', so when displayed, it will be little less nice, but likely more correct, e.g., /usr/bin/python (no version) instead of python3, following system configuration. Absolute path is prefered for subprocess execution by tools like bandit and sys.executable, even if we would not be using it here, would be an relevant example of absolute path usage. This will also make the code Python 4 ready and removes the need to handle Windows separately. sys.executable may be empty or None, but we already have a mechanism to set GRASS_PYTHON in special cases which is simply using GRASS_PYTHON externally, so I'm not adding any fallback for sys.executable being empty.