docs: Replace keywords index by MkDocs tags
This removes custom keywords index and replaces the keywords.html page by a MkDocs tags listing.

The default MkDocs tag anchor contains tag: prefix to help distinguishing tag headings from others, but we likely won't have a heading of the same name on the same page. However, we were using anchors without any prefix, so this will make most of the anchor links work (if there are any external ones anywhere). It also makes the anchor little simple to read and write if managed manually for any reason.

The check for anchor existence is done by MkDocs before the tags page is generated. This creates false positive for the Keywords section (to be removed) and for topics. To avoid this warning for topics, and perhaps to be little more helpful, instead of linking to the keyword, this new code uses MkDocs tags listing to list pages with that keyword right on the topic page.

This disables custom keywords build and disables the addons keyword integration in the CI workflow. The Python code to generate the keywords in Markdown is still in place, but it can be removed in the future possibly together with the custom HTML code (they are mixed togeher).