docs: Add contributing doc to main doc
Add main Markdown files relevant to contributing, esp. style guide, to the MkDocs-based documentation. Link CONTRIBUTING.md file, but don't include it, because it is meant for GitHub directly.

Move investigating history from git to development.

Adjust the build script to work nicely with files which don't have a YAML header and don't have any generated header. (An empty YAML header is not hidden, but leaks to HTML as dashes.)

Change the mutliple sections with single link to a single list, and conceptualize it as a list of API references.