docs: Remove Keywords section from generated doc (#5444)
This removes the Keywords section from the generated documentation in favor of showing MkDocs generated tags.

This also adds keywords to metadata of some files replacing a hard-coded Keywords section and adding keywords so there is at least two keywords when keywords are present.

There are two major differences:

1. The MkDocs tags are alphabetically ordered and are at the top of the page as "bubbles" instead of priority ordered tags as text inside a section.

2. All the tags go to the same page with a list of tags and a list of pages with each tag. This is like the original keywords index just with bubbles instead of plain headings. There is no distinction made for the first keyword being a category (raster, vector, ...) - going to category index - and second keyword being a topic - going to a topic page.

This preserves the current topic functionality which depends on the presence of keywords which are now obtained from the metadata rather than from parsing the Markdown text. Unlike the tool categories (raster, vector, ...), the topics are not linked from the menu at this point.