docs: Separate topics from keywords by 2nd tags plugin
The additional change in #5446 adding tag listing for topic pages turned out to be harmful as it makes the tags (aka keywords) on top of pages link to the tag heading in the topic pages if the listing exists. While this sort of looks like the original custom HTML behavior of second keyword being a tag, this creates inconsitency with an seemingly arbitrary keyword (not just the second or 'topic' keyword) linking to a different page than the list of keywords, but not even to the topic page itself, but to a heading inside the page because instead of a a topic, the link goes to the keyword (tag) in the listing.

This preserves the additional list of tools for topic pages introduced in #5446 by creating another instance of the material tags plugin in MkDocs using the same front matter metadata item, but different listing directive name and disabling tags for pages. This introduces a second independent tag system so keywords and topics don't mix.

This is still easy to remove or modify if we decide to change or remove topics in the future.