r.series: Say pipe in addition to showing the symbol
At least some of ASCII symbols are better described by a word than just by showing the symbol itself. This way it is clear that what the symbol is while also being able to copy the symbol. When pipe is a parameter, we allow, even promote, using the word rather than the symbol itself.