lib/parser: Use a function to get num of items in a tuple
This uses a function to get the required number of items in a tuple if the parameter value is defined as a tuple using key_desc. This is a first step to providing this information in the API and tool metadata which is important because it has implications for type (tool option parameter of type int with key_desc with comma has to have the comma, so it is actually not an int in e.g. Python).

There is little more what can be done here and that's not doing the check when there is no tuple, but this would be better done with some tests in place for this parser feature.

The function is actually already in #5490 so this is a draft also for this reason, but the logic should be reviewable.