docs: Replace grass-stable URLs by relative URLs
The core and addons doc mostly use absolute paths to refer to each other. However, when the docs are build together, the links may point to a wrong version. This is now even a bigger issue with Sphinx-based Python doc which is more linked from the main doc than it was before.

Having the absolute URLs pointing to grass-stable is still a good solution. It is straigforward. It works locally and when each piece is build separately.

The new script and the build step in CI is done after the two documentations are put together and with anticipation that libpython will be build into it. The replacement is done on Markdown level, so MkDocs will complain that the libpython links are broken, but that does not break the build and hopefully having both MkDocs and Sphinx docs is something to change in the future. No replacements are done for links from libpython, but there are currently none. The only links which are there are relative already.