grass.script: Add separate function to sort through run_command style parameters
Parameters for the run_command family of functions are a mix of Popen parameters and Python-ified parameters of a tool. This refactors the splitting functionality into a separate function allowing for use without the actual Popen call which is part of start_command. This is useful whan you need to call Popen or subprocess.run by yourself, but still want the run_command syntax.