grass.app: Add internal CLI with agrparse subcommands
This adds a CLI to grass.app subpackage using `__main__.py` and Python argparse with subcommands. At this point, this is meant as internal and experimental. The original motivation for this is testing locking in #5443. The subcommands add a lot of flexibility in what can done with argparse supporting directly several use cases and being quite stable and standardized.

Just as an example (to have this separate from #5443 while functional), this adds two subcommands, help and man, which both end up calling g.manual.