r.sim: more accurate estimation of iterations with null cells in region (#5588)
This is a followup of #5583, the averages used for estimating several other variables, including number of iterations, were taking into account number of cells in the rectangular region even if there are null cells. So for example when masking for a subwatershed, the values would be less accurate, average velocity would be smaller and the time step is greater than it should be.