CQ: always include stdbool.h in C code
Always `#include <stdbool.h>` where `bool` is used.

For the same reason, revert (part of 1e0c11f526f64755f877957ce6c14b2159c3823f)
definition of `TRUE` and `FALSE` (in gis.h) to `1` and `0`, in order to
avoid the inclusion everywhere it may/should have been needed. In addition,
starting with C23 `bool` is not an int, but a type of its own, making it
not interchangeable with int.